+++ src/blocks/Config/validatorsType.ts
@@ -1,5 +1,5 @@
 import type { PasteScope } from '../../abstract/controllers/ClipboardController';
-import type { Metadata, MetadataCallback } from '../../types/index';
+import type { Metadata, MetadataCallback, Tags, TagsCallback } from '../../types/index';
 import { deserializeCsv } from '../../utils/comma-separated';
 import type { ModeCameraType } from '../CameraSource/constants';
 import { CameraSourceTypes } from '../CameraSource/constants';
@@ -78,6 +78,17 @@ const asMetadata = (value: unknown): Metadata | MetadataCallback => {
   throw new Error('Invalid metadata value. Must be an object or function.');
 };
 
+const asTags = (value: unknown): Tags | TagsCallback => {
+  if (Array.isArray(value)) {
+    return value as Tags;
+  }
+  if (typeof value === 'function') {
+    return value as TagsCallback;
+  }
+
+  throw new Error('Invalid tags value. Must be an array or function.');
+};
+
 const asObject = <T>(value: unknown): T => {
   if (typeof value === 'object') {
     return value as T;
@@ -114,7 +125,7 @@ const asFilesViewMode = (value: unknown): FilesViewMode => {
 
 const asDynamicButtonViewMode = (value: unknown): DynamicButtonMode => {
   const strValue = asString(value);
-  if (['auto', 'menu', 'toolbar', 'compact'].includes(strValue)) {
+  if (['auto', 'menu', 'toolbar', 'compact', 'plain'].includes(strValue)) {
     return strValue as DynamicButtonMode;
   }
   throw new Error(`Invalid value: "${strValue}"`);
@@ -133,6 +144,7 @@ export {
   asObject,
   asStore,
   asString,
+  asTags,
   asPasteScope,
   asDynamicButtonViewMode,
 };