+++ src/types/index.ts
@@ -1,4 +1,5 @@
 export type { CustomConfig, CustomConfigDefinition } from '../abstract/customConfigOptions';
+export type { CustomLocaleDefinition } from '../abstract/localeRegistry';
 export type * from '../abstract/managers/plugin/PluginTypes';
 export type { ActivityType, CustomActivities } from '../lit/activity-constants';
 export type * from './events';